07--project/requirements.txt
tqdm
notebook==7.1.2
google-generativeai
elasticsearch
pandas
scikit-learn
openai
streamlit
python-docx
jupyter
sentence-transformers
python-dotenv
psycopg2
Flask